+++ tests/testthat/test_decisions.R
@@ -96,3 +96,30 @@ test_that("add_semicontinuous_decisions (solve)", {
   expect_true(isTRUE(all(round(na.omit(values(s)), 5) <= 0.3)))
   expect_true(isTRUE(all(na.omit(values(s)) >= 0)))
 })
+
+test_that("add_semicontinuous_decisions (solve with locked constraints)", {
+  skip_on_cran()
+  skip_if_not(any_solvers_installed())
+  # generate solution
+  data(sim_pu_raster, sim_features)
+  s <- problem(sim_pu_raster, sim_features) %>%
+    add_min_set_objective() %>%
+    add_relative_targets(0.1) %>%
+    add_locked_in_constraints(sim_locked_in_raster) %>%
+    add_locked_out_constraints(sim_locked_out_raster) %>%
+    add_semicontinuous_decisions(0.3) %>%
+    add_default_solver(time_limit = 5) %>%
+    solve()
+  s2 <- problem(sim_pu_raster, sim_features) %>%
+    add_min_set_objective() %>%
+    add_relative_targets(0.1) %>%
+    add_locked_out_constraints(sim_locked_out_raster) %>%
+    add_locked_in_constraints(sim_locked_in_raster) %>%
+    add_semicontinuous_decisions(0.3) %>%
+    add_default_solver(time_limit = 5) %>%
+    solve()
+  # check that solutions have correct decisions
+  expect_true(isTRUE(all(round(na.omit(values(s)), 5) <= 0.3)))
+  expect_true(isTRUE(all(na.omit(values(s)) >= 0)))
+  expect_equal(values(s), values(s2))
+})